requirements.txt
litellm>=1.0.0
fastapi>=0.95.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
pydantic>=1.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
alembic>=1.7.7
psycopg2-binary>=2.9.3
requests>=2.28.1
python-slugify>=6.1.2
python-dateutil>=2.8.2
uvicorn>=0.21.0
